requirements.txt
apache-airflow[postgres]==2.5.2
pandas==1.5.2
pyarrow==12.0.0
pydantic==1.10.7
pytest==7.2.0
sqlalchemy==1.4.48